even more improve exception (#1182)
