+++ src/revChatGPT/V0.py
@@ -83,13 +83,13 @@ class Chatbot:
         user: str = "User",
     ) -> dict:
         if completion.get("choices") is None:
-            error = t.ChatbotError("ChatGPT API returned no choices")
+            error = t.ResponseError("ChatGPT API returned no choices")
             raise error
         if len(completion["choices"]) == 0:
-            error = t.ChatbotError("ChatGPT API returned no choices")
+            error = t.ResponseError("ChatGPT API returned no choices")
             raise error
         if completion["choices"][0].get("text") is None:
-            error = t.ChatbotError("ChatGPT API returned no text")
+            error = t.ResponseError("ChatGPT API returned no text")
             raise error
         completion["choices"][0]["text"] = remove_suffix(
             completion["choices"][0]["text"],
@@ -115,15 +115,15 @@ class Chatbot:
         full_response = ""
         for response in completion:
             if response.get("choices") is None:
-                error = t.ChatbotError("ChatGPT API returned no choices")
+                error = t.ResponseError("ChatGPT API returned no choices")
                 raise error
             if len(response["choices"]) == 0:
-                error = t.ChatbotError("ChatGPT API returned no choices")
+                error = t.ResponseError("ChatGPT API returned no choices")
                 raise error
             if response["choices"][0].get("finish_details") is not None:
                 break
             if response["choices"][0].get("text") is None:
-                error = t.ChatbotError("ChatGPT API returned no text")
+                error = t.ResponseError("ChatGPT API returned no text")
                 raise error
             if response["choices"][0]["text"] == "<|im_end|>":
                 break
